% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_input_format.R
\name{describe_input_format}
\alias{describe_input_format}
\title{Mostra exemplo do formato de entrada esperado}
\usage{
describe_input_format()
}
\value{
Invisivelmente NULL. Apenas imprime no console.
}
\description{
Imprime no console um exemplo simples do formato aceitável para ser lido
por `filter_energy()` e funções relacionadas.
}
\examples{
\dontrun{
describe_input_format()
}
}
