% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRHeat_PACKAGE.R
\docType{package}
\name{miRHeat-package}
\alias{miRHeat}
\alias{miRHeat-package}
\title{miRHeat: Ferramentas para análise de interações miRNA–UTR}
\description{
O pacote miRHeat fornece funções para ler arquivos de predição de interação
entre genes e miRNAs,
selecionar métricas de score, preparar dados e gerar heatmaps
de interações miRNA × UTR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bianicoletti/miRHeat}
  \item Report bugs at \url{https://github.com/bianicoletti/miRHeat/issues}
}

}
\author{
\strong{Maintainer}: Maria Nicoletti \email{maria.nicoletti@unesp.br}

}
\keyword{internal}
