% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{select_score}
\alias{select_score}
\title{Select or redefine Score column}
\usage{
select_score(df, score_column = NULL)
}
\arguments{
\item{df}{A data.frame containing at least `miRNA` and `target` columns.}

\item{score_column}{Character or NULL. Name of the numeric column to be used as Score.}
}
\value{
A data.frame with a `Score` column.
}
\description{
Allows the user to define or override the `Score` column in a data.frame.
This is mainly useful for custom or external tabular data where the Score
was not automatically detected.
}
\examples{
df <- data.frame(miRNA = c("a","b"), target = c("T1","T2"), energy = c(-10, -20))
select_score(df, "energy")

}
