% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRHeat_PACKAGE.R
\docType{package}
\name{miRHeat-package}
\alias{miRHeat}
\alias{miRHeat-package}
\title{miRHeat: Tools for analyzing miRNA–UTR interactions}
\description{
The miRHeat package provides functions for reading interaction prediction files
between genes and miRNAs,
selecting score metrics, preparing data, and generating heatmaps
of miRNA × UTR interactions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bianicoletti/miRHeat}
  \item Report bugs at \url{https://github.com/bianicoletti/miRHeat/issues}
}

}
\author{
\strong{Maintainer}: Maria Nicoletti \email{maria.nicoletti@unesp.br}

}
\keyword{internal}
