% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_mirheat}
\alias{plot_mirheat}
\title{Generate miRNA vs UTR interaction heatmap}
\usage{
plot_mirheat(df, output_file = NULL, width = 2000, height = 1800)
}
\arguments{
\item{df}{Data.frame containing the columns miRNA, UTR (or utr), and Score.}

\item{output_file}{Optional path to export the PNG file (default = NULL).}

\item{width}{Width of the PNG when exporting (in pixels).}

\item{height}{Height of the PNG when exporting (in pixels).}
}
\value{
Invisibly returns the ComplexHeatmap object.
}
\description{
Generates a heatmap of interaction scores between miRNAs and UTRs.
Automatically determines the optimal number of clusters using
dynamic tree cutting.
}
\examples{
\dontrun{
file <- system.file("extdata", "exemplo_blocos.txt", package = "miRHeat")
df <- parse_file(file)
df <- select_score(df, "interaction_energy")
df <- prepare_for_heatmap(df)
plot_mirheat(df)
}
}
