% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_and_prep.R
\name{prepare_for_heatmap}
\alias{prepare_for_heatmap}
\title{Prepara tabela padronizada para heatmap}
\usage{
prepare_for_heatmap(df)
}
\arguments{
\item{df}{Data.frame com pelo menos miRNA, target/gene e Score.}
}
\value{
Data.frame com colunas ordenadas: miRNA, Target, Gene, UTR, Score, e outras colunas opcionais.
}
\description{
Recebe o data.frame com Score e cria/garante as colunas:
miRNA, Target, Gene, UTR, Score. Retorna um data.frame pronto para o heatmap.
}
\examples{
\dontrun{
df_ready <- prepare_for_heatmap(df_filtered)
}
}
