% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_for_heatmap}
\alias{prepare_for_heatmap}
\title{Prepara tabela padronizada para heatmap}
\usage{
prepare_for_heatmap(df)
}
\arguments{
\item{df}{Data.frame com pelo menos miRNA, target/gene e Score.}
}
\value{
Data.frame com colunas ordenadas: miRNA, target, gene, utr, Score e outras colunas opcionais.
}
\description{
Recebe o data.frame com Score e cria/garante as colunas:
miRNA, target, gene, utr, Score. Retorna um data.frame pronto para o heatmap.
}
\examples{
df <- data.frame(miRNA = c("a","b"), target = c("G1_UTR_1","G2"), Score = c(-10, -20))
prepare_for_heatmap(df)
}
