% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_and_prep.R
\name{select_score}
\alias{select_score}
\title{Seleciona qual coluna numérica será usada como Score}
\usage{
select_score(df, score_column = NULL)
}
\description{
Seleciona qual coluna numérica será usada como Score
}
